Fix compiler warnings

+++ src/ArchiveFormats/ElfFile.h
@@ -61,7 +61,7 @@ public:
     mFilePath = filePath;
     auto [buffer, size] =
         read_file_to_buffer<unsigned char>(filePath.string().c_str());
-    if (size < sizeof(Elf32_Header)) {
+    if ((ulong)size < sizeof(Elf32_Header)) {
       return;
     }
     mFileStream = buffer;
